checkWinner and reset functions
I created the checkWinner and the reset functions in the Gameboard and the Square components. i modified the style of the Gameboard. I created the resetContext function to set the context to the inital state and navigate back to the home component. i changed the favicon and the logos in the public folder